src/requirements.txt
uvicorn
fastapi
requests
validators
pymongo
python-multipart
python-jose
passlib
pycountry
python-dateutil